Tinyml_model/requirements.txt
numpy
pandas
scikit-learn
tensorflow
joblibcls